chore: Refactor Toc.vue component to improve performance and readability
